Topic service added to create, get topics from the kafka cluster

+++ app-data-config/src/main/java/com/ruminderhub/twitter_kafka/config/RetryConfigData.java
@@ -11,7 +11,7 @@ public class RetryConfigData {
 
     private Long initialIntervalMs;
     private Long maxIntervalMs;
-    private Double multiplier;
+    private Integer multiplier;
     private Integer maxAttempts;
     private Long sleepTimeMs;
 }
